+++ src/Textil.java
@@ -3,7 +3,7 @@ import java.util.Objects;
 public class Textil extends Producte implements Comparable<Textil> {
     private String composicio;
 
-    public Textil(double preu, String nom, String codiBarres, String composicio) {
+    public Textil(float preu, String nom, String codiBarres, String composicio) {
         super(preu, nom, codiBarres);
         this.composicio = composicio;
     }
